+++ changelog.md
@@ -2,6 +2,8 @@
 
 ### v1.0.0-rc.4
 
+- Added a warning for unloaded changes to the `application.ex` file when using Live reload.
+
 - Fixed a bug where calling `Tamnoon.MethodManager.trigger_method/2` inside the `:initial_state` callback would not trigger the method properly.
 
 - Fixed a bug where changing a `<video>`'s `src` attribute would not reload the video.